+++ src/resourcePackage/Resource.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 abstract public class Resource
 {
-	private int _localId;
+	private long _localId;
 	private URL _url;
 	private File _file;
 	private String _name;
@@ -19,7 +19,7 @@ abstract public class Resource
 	
 	final public static int UNDEFINED = -1;
 	
-	public Resource(int localId, URL url, File file, String name, List<Author> authors)
+	public Resource(long localId, URL url, File file, String name, List<Author> authors)
 	{
 		setlocalId(localId);
 		setUrl(url);
@@ -34,7 +34,7 @@ abstract public class Resource
 		this(UNDEFINED, url, file, name, authors);
 	}
 	
-	public Resource(int localId, URL url, File file, String name, Author author)
+	public Resource(long localId, URL url, File file, String name, Author author)
 	{
 		this(localId, url, file, name, new ArrayList<Author>());
 		getAuthors().add(author);
@@ -45,7 +45,7 @@ abstract public class Resource
 		this(UNDEFINED, url, file, name, author);
 	}
 	
-	public Resource(int localId, URL url, File file, String name)
+	public Resource(long localId, URL url, File file, String name)
 	{
 		this(localId, url, file, name, new ArrayList<>());
 	}
@@ -74,12 +74,12 @@ abstract public class Resource
 		return getName();
 	}
 
-	public int getlocalId()
+	public long getlocalId()
 	{
 		return _localId;
 	}
 
-	public void setlocalId(int localId)
+	public void setlocalId(long localId)
 	{
 		_localId = localId;
 	}
